updated all effects
