--- cameraview/src/main/java/com/otaliastudios/cameraview/shadereffects/effects/NoEffect.java
@@ -1,42 +1,0 @@
-package com.otaliastudios.cameraview.shadereffects.effects;
-
-import com.otaliastudios.cameraview.shadereffects.BaseShaderEffect;
-
-/**
- * Displays the normal preview without any effect.
- */
-public class NoEffect extends BaseShaderEffect {
-    /**
-     * Initialize
-     */
-    public NoEffect() {
-    }
-
-    @Override
-    public String getFragmentShader() {
-
-        /*String shader = "#extension GL_OES_EGL_image_external : require\n"
-                + "precision mediump float;\n"
-                + "varying vec2 vTextureCoord;\n"
-                + "uniform samplerExternalOES sTexture;\n"
-                + "void main() {\n"
-                + "  gl_FragColor = texture2D(sTexture, vTextureCoord);\n"
-                + "}\n";*/
-
-        //This fragment shader is copied from black and white,
-        //this is for testing purpose and need to be removed once testing is done.
-        String shader = "#extension GL_OES_EGL_image_external : require\n"
-                + "precision mediump float;\n"
-                + "varying vec2 vTextureCoord;\n"
-                + "uniform samplerExternalOES sTexture;\n" + "void main() {\n"
-                + "  vec4 color = texture2D(sTexture, vTextureCoord);\n"
-                + "  float colorR = (color.r + color.g + color.b) / 3.0;\n"
-                + "  float colorG = (color.r + color.g + color.b) / 3.0;\n"
-                + "  float colorB = (color.r + color.g + color.b) / 3.0;\n"
-                + "  gl_FragColor = vec4(colorR, colorG, colorB, color.a);\n"
-                + "}\n";
-
-        return shader;
-
-    }
-}
