+++ cameraview/src/main/java/com/otaliastudios/cameraview/shadereffects/effects/NoFilterEffect.java
@@ -1,0 +1,11 @@
+package com.otaliastudios.cameraview.shadereffects.effects;
+
+import com.otaliastudios.cameraview.shadereffects.BaseShaderEffect;
+
+public class NoFilterEffect extends BaseShaderEffect {
+
+    @Override
+    public String getFragmentShader() {
+        return mFragmentShader;
+    }
+}
